SKA-475: Add label support
By default, the FMU Importer does not use any labels when creating Publish / Subscribe services.
Labels can be set by configuring PublisherInstance, SubscriberInstance, and Namespace in the root of the configuration file.
These labels allow to limit the communication between services with the same namespace / instance name (e.g., if multiple instances of the same FMU are participating in a simulation).